main

+++ base/pipelines/fine_tuning.py
@@ -569,7 +569,7 @@ def train_lora_model(data, video_folder, args):
                 last_hidden_state = outputs.hidden_states[-1].to(torch.float16)
                 print(f"train_lora_model last_hidden_state shape: {last_hidden_state.shape}, dtype: {last_hidden_state.dtype}") #[1, 50, 768] torch.float16
                 
-                cross_attention = CrossAttentionTI(embed_dim=768, num_heads=8).to(unet.device)
+                cross_attention = CrossAttentionTI(embed_dim=768, num_heads=8).to(unet.device).to(torch.float16)
                 
                 # Trasponiamo le dimensioni per adattarsi al MultiheadAttention
                 text_features = text_features.transpose(0, 1)
